+++ CriteoPublisherSdk/Sources/Network/CR_ApiQueryKeys.m
@@ -114,9 +114,6 @@
 + (NSString *)trackingAuthorizationStatus {
   return @"trackingAuthorizationStatus";
 }
-+ (NSString *)mopubConsent {
-  return @"mopubConsent";
-}
 + (NSString *)profileId {
   return @"profileId";
 }